requirements.txt
groq>=0.4.0
opencv-python>=4.8.0
numpy>=1.24.0
flask>=2.3.0
werkzeug>=2.3.0
fpdf2>=2.7.9

